+++ CHANGELOG.md
@@ -35,6 +35,7 @@ The following changes have been completed, and will be included in the next tagg
   - added `isHRefPartsWithHash()` type guard
   - added `isHRefPartsWithHostname()` type guard
   - added `isHRefPartsWithPathname()` type guard
+  - added `isPRHRefParts()` type guard
 * ParsedURL
   - added `ParsedURL` interface
 * URL